Fix Dockerfile pip install
