+++ docker-entrypoint.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+service postgresql start
+exec "$@"
